feat: restructure customer views and update routing for product and registration components
