+++ src/views/customer/CustomerProductsView.vue
