+++ src/views/customer/ProductDetailsView.vue
